fix tx performance with multiple lcores
Find and fixed the issue with not having multiple TX queues when the
number of TX lcores is more then one.
Fixed the issue with stats when the device does not support multiple
queues like i40e.
Fixed the problem with counting too many TX packets by mistake.

Still a WiP to find the tx performance with more then two lcores.

Signed-off-by: Keith Wiles <keith.wiles@intel.com>